EVIL SQUIDWART
This is the first virus I've ever made for Android, heavily inspired by My Horror.
This was made with Sketchware Pro within a half hour, so this is pretty low effort. You'll see that as a pattern for my future viruses.
THIS IS NOT DESTRUCTIVE! This simply changes your wallpaper and plays music. To stop the music, close the app window.
Change your wallpaper back if you want.